requirements.txt
streamlit
scikit-learn
matplotlib
shap
xgboost
lightgbm
optuna
openai==0.28
pandas
numpy
python-dotenv
seaborn